Facilitate deployments to heroku
Heroku uses 'buildpacks' which contain tools needed to build an
application for a given technology, but allow little customization
and have a pre-set number of tools included (or I haven't found a
way to include both `composer` and `node` in 1 environment, short
of defining such a build environment by hand).

This is why we need to invoke 2 seperate buildpacks sequentially:

1. `composer install``
2. `npm run build`

Defining the environment variables like this allows the developer to
easily fill in the required values with minimal overhead.

Heroku heavily uses environment variables to configure applications,
and doesn't allow for after-deployment adjustments.

Using environment variables for configuration does not fit into
the current way the config is set up since this uses constants,
which need to be fixed at the start of the program.

This leads us to the following 2 points:

1. Since these are constants, they cannot be read directly from the
environment and we have to perform an in-file replace to make them
active.
2. A change in environment variables triggers a restart of a dyno,
but not a redeploy. This means that we cannot configure these at
build, so we have to perform this change when a Dyno is (re)started.

The [.profile](https://devcenter.heroku.com/articles/dynos#the-profile-file)
file allows us to manipulate the environment and files for the Dyno,
or runner, allowing us to alter the `config.php` file and apply
the desired config, while still sticking to non-mutable configuration
once the application is booted..

We also need a custom logger to propagate the CodeIgniter logs
towards Heroku.

See https://devcenter.heroku.com/articles/php-logging#codeigniter-3-x